moved purpose models from commons to its package
